K8S-lab14/backend/requirements.txt
Flask==2.2.3
mysql-connector-python==8.0.31
Werkzeug==2.2.3  # Downgrade Werkzeug to match Flask version
